backend/aws_requirements.txt
# Core API requirements
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# AWS Lambda adapter
mangum==0.17.0

# AWS SDK
boto3==1.34.0
botocore==1.34.0

# Data processing (keep existing dependencies)
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
scipy==1.11.1

# Optional: Database connectors
# Uncomment based on your database choice:
# psycopg2-binary==2.9.7  # PostgreSQL
# pymongo==4.5.0          # MongoDB 